Adding datViz functions and updating readme to for datViz use
