+++ countLikes/README.md
@@ -23,13 +23,17 @@ After installing the supporting Libraries, fill out the "_/input/input.csv_" fil
 - **ID:** ID of the Interested element.
     - Photos: "http://...facebook.com/photo.php?fbid=**PHOTO FACEBOOK ID**&set.."
     - Pages: Use http://findmyfbid.com/
-    - Post: Click on TimeStamp of the Post. You will be navigated to a url of this format: "https://www.facebook.com/kavilovesyou/posts/**POST FACEBOOK ID**"
+    - Post: Click on TimeStamp of the Post. You will be navigated to a url of this format: "https://www.facebook.com//posts/**POST FACEBOOK ID**"
 - **Type:** What type of element is it i.e. "Photo","Page","Post"
 
 ## Execute
-    $>python countLikes.py <input file path> <output file path>
+    $>python countLikes.py <input file path> <output file path> <datViz>
 
-"input file path" = Path to your input you created in Step 2 above.    
+- "input file path" = Path to your input you created in Step 2 above.    
+- "output file path" = Path to the output.
+- "datViz" = Visual reprsentation. Please see below for visuals supported.
+	- Bar: See Bar graph of all Names and corresponding # of Likes.
+	- More Coming Soon!
 You will find your "output CSV" in the _output_ directory unless stated otherwise.
 
 ## Resources
