+++ countLikes/countLikes.py
@@ -1,65 +1,32 @@
 #!/usr/bin/env python
 #title          : countLikes.py
-#description    : Take CSV with (Name,Facebook_ID,Type(Picture,Page)) and outputs a CSV with Total Like Counts using Graph API
+#description    : Take CSV with (Name,Facebook_ID,Type) and outputs a CSV with Total Like Counts using Graph API
 #author         : @asinghal 
 #=======================================================================================================================
-
 import sys
-import requests
-import json
 import pandas as pd
 import csv
 from numpy import array,arange
-from matplotlib import pyplot as plt
-from matplotlib import style
-style.use('ggplot')
+from lib.datViz import *
+from lib.parse import *
+
 # =================================================
-### Helper Functions
+### HELPER FUNCTION
 # =================================================
-
-def getDat(url,token):
-	param={'access_token': token}
-	r=requests.get(url,params=param)
-	dat=r.text
-	dat=json.loads(dat)
-	return dat
-
-def getUrl(fbookid, goal):
-	url= 'Invalid entry for "goal" or "id"'
-	if goal=='page':
-		url= 'https://graph.facebook.com/%d/?fields=likes' %fbookid
-	else:
-		url= 'https://graph.facebook.com/%d/likes?summary=1' %fbookid
-	return url
-
-def getObjectLikes(json_input):
-	return json_input['summary']['total_count']
-
-def getPageLikes(json_input):
-	return json_input['likes']
 
 def getLikes(fbookid,token,goal):
+	"""Return Like count for a given ID + Type.
+	
+	fbookid -- Object Facebook ID
+	token -- Access Token. See GRAPH API Documentation to obtain your accesstoken.
+	goal -- Type of Object. See Readme for supported types.
+	"""
 	url=getUrl(fbookid, goal)
-	json= getDat(url,token)
-	if goal=='page':
+	json=getDat(url,token)
+	if isPage(goal):
 		return getPageLikes(json)
 	else:
 		return getObjectLikes(json)
-
-def getbar(dictionary):
-	l=arange(len(dictionary))
-	plt.bar(l,dictionary.values(), color='#ff6701', align='center')
-	plt.ylabel('# of Likes')
-	plt.xticks(l,dictionary.keys())
-	plt.xlabel('Name')
-	plt.savefig("output/bar.png")
-	plt.show()
-
-def getLikesDict(arr):
-	d=dict()
-	for item in arr:
-		d[item[0]]=item[2]
-	return d
 
 # =================================================
 ### LIKE COUNTER
@@ -68,8 +35,8 @@ def getLikesDict(arr):
 def run_counter(infile,outfile,token):
 	"""Takes in a CSV of Facebook IDs and types to give Facebook Like Counts.
 	
-	infile -- CSV Input (Facebook_ID, Type)
-	outfile -- CSV Output of (Facebook_ID,Like Counts,Type)
+	infile -- CSV Input. See Readme for Format.
+	outfile -- CSV Output.
 	"""
 	x=pd.read_csv('%s' %infile , delimiter=",")
 	arr=array(x)
@@ -77,7 +44,7 @@ def run_counter(infile,outfile,token):
 	with open('%s' %outfile, 'a') as h:
 		ff = csv.writer(h, delimiter=',')
 		for i in arr:
-			if cnt<5:
+			if cnt<10:
 				print "Only %d more left!" %cnt
 			name=i[0]
 			facebook_id=i[1]
@@ -95,10 +62,10 @@ def run_counter(infile,outfile,token):
 ### Main
 # =================================================
 from argparse import ArgumentParser
-parser=ArgumentParser(usage="python countLikes.py input_filename output_filename [--viz]")
+parser=ArgumentParser(usage="python countLikes.py input_filename output_filename datViz")
 parser.add_argument('infile', help='Input CSV Data File Path')
 parser.add_argument('outfile', help='Output CSV File')
-parser.add_argument('viz',default="None", help='Viz Type')
+parser.add_argument('viz',default="bar", help='Viz Type')
 ar=parser.parse_args()
 
 infile='input/'+ar.infile
@@ -119,6 +86,7 @@ if datViz=="bar":
 	outarr=array(out)
 	out_dict=getLikesDict(outarr)
 	getbar(out_dict)
+	print ("Thanks for Using countLikes!")
 else:
-	print ("Viz " + datViz + " Not Supported")
+	print ("DatViz not supported/specified!")
